requirements.txt
Django==1.10.3
requests==2.11.1
hashids==1.1.0
djangorestframework==3.5.3
